+++ src/ClockListener.java
@@ -1,4 +1,4 @@
 
 public interface ClockListener {
-	public void notificar(int i);
+	public void notificar(long i);
 }
